Add removes to ProgramYear
Added DirectorsEntity and CompetenciesEntity trait and applied them
where appropriate.
