+++ src/Ilios/CoreBundle/Entity/School.php
@@ -5,6 +5,7 @@ namespace Ilios\CoreBundle\Entity;
 use Doctrine\ORM\Mapping as ORM;
 
 use Ilios\CoreBundle\Traits\AlertableEntity;
+use Ilios\CoreBundle\Traits\CompetenciesEntity;
 use Ilios\CoreBundle\Traits\IdentifiableEntity;
 use Ilios\CoreBundle\Traits\InstructorGroupsEntity;
 use Ilios\CoreBundle\Traits\SessionTypesEntity;
@@ -44,6 +45,7 @@ class School implements SchoolInterface
     use AlertableEntity;
     use SessionTypesEntity;
     use InstructorGroupsEntity;
+    use CompetenciesEntity;
 
     /**
      * @var int
@@ -315,34 +317,6 @@ class School implements SchoolInterface
             $this->alerts->add($alert);
             $alert->addRecipient($this);
         }
-    }
-
-    /**
-     * @param Collection $competencies
-     */
-    public function setCompetencies(Collection $competencies)
-    {
-        $this->competencies = new ArrayCollection();
-
-        foreach ($competencies as $competency) {
-            $this->addCompetency($competency);
-        }
-    }
-
-    /**
-     * @param CompetencyInterface $competency
-     */
-    public function addCompetency(CompetencyInterface $competency)
-    {
-        $this->competencies->add($competency);
-    }
-
-    /**
-     * @return ArrayCollection|CompetencyInterface[]
-     */
-    public function getCompetencies()
-    {
-        return $this->competencies;
     }
 
     /**
